feat: 모달 컴포넌트 리팩토링 및 새로운 모달 추가
