--- src/components/Modals/DeleteAccountForm.tsx
@@ -1,67 +1,0 @@
-import { useState } from 'react';
-
-import Button from '../atoms/Button/Button';
-import Input from '../atoms/Input/Input';
-
-interface DeleteAccountFormProps {
-  onSuccess: () => void;
-}
-
-const DeleteAccountForm = ({ onSuccess }: DeleteAccountFormProps) => {
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
-  const [success, setSuccess] = useState(false);
-
-  const handleDelete = async () => {
-    setError('');
-    try {
-      const res = await fetch('/users/me', {
-        method: 'DELETE',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ password })
-      });
-      const data = await res.json();
-      if (data.success) {
-        setSuccess(true);
-        setTimeout(() => {
-          onSuccess();
-          setSuccess(false);
-        }, 5000);
-      } else {
-        setError(data.error || '서버 오류로 실패하였습니다.');
-      }
-    } catch {
-      setError('서버 오류로 실패하였습니다.');
-    }
-  };
-
-  if (success) {
-    return (
-      <div style={{ padding: 24, textAlign: 'center' }}>
-        정보 변경이 완료되었습니다.
-        <br />
-        5초 후 창이 닫힙니다.
-      </div>
-    );
-  }
-
-  return (
-    <div style={{ padding: 24 }}>
-      <label htmlFor="delete-password">비밀번호 확인</label>
-      <Input
-        id="delete-password"
-        type="password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-        placeholder="비밀번호를 입력하세요"
-        variant="long"
-      />
-      <Button variant="account" onClick={handleDelete} style={{ marginTop: 16 }}>
-        회원 탈퇴
-      </Button>
-      {error && <div style={{ color: 'red', marginTop: 12 }}>{error}</div>}
-    </div>
-  );
-};
-
-export default DeleteAccountForm;
